Move test entities.
