+++ test/Tests.Infrastructure/GraphSon/Entities/Bug_1884_Enum.cs
@@ -1,0 +1,8 @@
+﻿namespace ExRam.Gremlinq.Tests.Infrastructure.GraphSon.Entities
+{
+    public enum Bug_1884_Enum
+    {
+        ValueA,
+        ValueB
+    }
+}
