+++ test/Tests.Infrastructure/GraphSon/Entities/ClassWithField.cs
@@ -1,0 +1,7 @@
+﻿namespace ExRam.Gremlinq.Tests.Infrastructure.GraphSon.Entities
+{
+    public class ClassWithField
+    {
+        public string? Property;
+    }
+}
