+++ test/Tests.Infrastructure/GraphSon/Entities/ClassWithInitProperty.cs
@@ -1,0 +1,7 @@
+﻿namespace ExRam.Gremlinq.Tests.Infrastructure.GraphSon.Entities
+{
+    public class ClassWithInitProperty
+    {
+        public string? Property { get; init; }
+    }
+}
